Publish the GeoExt-pkg, it's advertized in catalog
On the gh-pages branch, we publish the locally created package BasiGX
which itself depends on the GeoExt package. Since in order to create
the BasiGX package, we had to add the GeoExt package, our local catalog
(`catalog.json` in the repo folder) also contains the description of
the GeoExt package.

When someone now tries to use the BasiGX package, that catalog is
queried. Since it also contains a reference to GeoExt (we simply
copied it over), the sencha cmd tries to resolve any GeoExt
dependencies from the BasiGX folder.

This commit ensures that we also publish the GeoExt package to the
BasiGX repo. In the future we may want to change the way we advertize
our catalog. See the TODO in the `update-gh-pages.sh`-file.