Make source & behaviour of selection configurable
This introduces two new configuration options for the hover plugin:

* `selectMulti`, and
* `selectEventOrigin`

With `selectMulti` you can control whether the select interaction may ever
possibly select more than one feature from the hover layer. The default is
`true` for backwards compatibillity.

With `selectEventOrigin` you can control if you want the origin of the
click event firing to be the collection of the select interaction (`add`
event, also the default for historical reasons) or the interaction itself
(its `select` event).