Be more cautious when creating a tilegrid
Since 71e66d5ee6ef7bd70f94538f998ff3df3fb3eeba we overwrite a layers `tileGrid`
configuration without checking if …

a) … such a `tileGrid` was already defined in the configuration
b) … the needed inputs for the generation are defined in the `mapConfig`

For a) we should respect what we were given, and for b) we should check if
the key `resolutions` exists in `appContext.mapConfig`.

Additionally we should really be logging that something unexpected happened in
the method `convertStringToNumericArray` instead of silently returning the
input.

This is a follow-up for PR #62.